+++ examples/machine-controller.yaml
@@ -536,7 +536,7 @@ rules:
   - "list"
   - "get"
   - "watch"
-# storageAttachments permissions are needed in vSphere to solve an the following issue: https://github.com/kubermatic/machine-controller/issues/1189
+# volumeAttachments permissions are needed by clusters using external CSI 
 - apiGroups:
   - "storage.k8s.io"
   resources:
